+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "flatbush"
+name = "geo-index"
 version = "0.1.0"
 edition = "2021"
 rust-version = "1.75"
