+++ src/kdtree/error.rs
