+++ src/rtree/error.rs
