+++ src/rtree/sort/str.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "rayon")]
 use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterator};
 
-use crate::flatbush::sort::{Sort, SortParams};
+use crate::rtree::sort::{Sort, SortParams};
 use crate::indices::MutableIndices;
 use crate::r#type::IndexableNum;
