+++ src/rtree/traversal.rs
